# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: VitalSub API
  version: 1.0.0
  description: General-purpose intelligent ingredient substitution API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recipe
    description: Recipe analysis and substitution
  - name: ingredients
    description: Ingredient data operations
  - name: feedback
    description: Substitution feedback

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyze-recipe:
    post:
      operationId: analyzeRecipe
      tags: [recipe]
      summary: Analyze a recipe and suggest substitutes
      description: |
        Parses the ingredient list, normalizes variants (e.g. "brown sugar" → "sugar"),
        then runs a 3-layer substitution pipeline: direct DB mappings, functional/category
        matches, and cosine nutrition similarity. Returns ranked substitutes scored by
        nutritional similarity, functional role, health improvement, and user feedback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRecipeRequest"
      responses:
        "200":
          description: Substitution analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeRecipeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List all ingredients
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"

  /ingredients/{name}/substitutes:
    get:
      operationId: getSubstitutes
      tags: [ingredients]
      summary: Get ranked substitutes for a named ingredient
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ranked list of substitutes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubstituteWithNutrition"

  /ingredients/{name}/nutrition:
    get:
      operationId: getIngredientNutrition
      tags: [ingredients]
      summary: Get nutrition data for an ingredient
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Nutrition data per 100g
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionData"
        "404":
          description: Ingredient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit user feedback for a substitution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "201":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    AnalyzeRecipeRequest:
      type: object
      properties:
        recipe:
          type: string
          description: Comma-separated ingredient list, e.g. "2 cups brown sugar, 1 cup milk"
      required: [recipe]

    NutritionData:
      type: object
      properties:
        calories: { type: ["number", "null"] }
        protein:  { type: ["number", "null"] }
        carbs:    { type: ["number", "null"] }
        fat:      { type: ["number", "null"] }
        sugar:    { type: ["number", "null"] }
        sodium:   { type: ["number", "null"] }

    SubstituteWithNutrition:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
          description: Composite score 0–100
        reason:
          type: string
          description: Why this substitute was recommended
        nutrition:
          $ref: "#/components/schemas/NutritionData"
      required: [name, score]

    IngredientResult:
      type: object
      properties:
        name:
          type: string
        normalizedName:
          type: string
          description: Canonical ingredient name if input was a recognized variant
        hasSubstitutes:
          type: boolean
        substitutes:
          type: array
          items:
            $ref: "#/components/schemas/SubstituteWithNutrition"
        originalNutrition:
          $ref: "#/components/schemas/NutritionData"
      required: [name, hasSubstitutes, substitutes]

    AnalyzeRecipeResponse:
      type: object
      properties:
        parsedIngredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientResult"
        originalNutritionTotal:
          $ref: "#/components/schemas/NutritionData"
        substitutedNutritionTotal:
          $ref: "#/components/schemas/NutritionData"
        substituteCount:
          type: integer
          description: Number of ingredients that have at least one substitute recommendation
      required:
        - parsedIngredients
        - originalNutritionTotal
        - substitutedNutritionTotal
        - substituteCount

    Ingredient:
      type: object
      properties:
        id:       { type: integer }
        name:     { type: string }
        category: { type: string }
      required: [id, name, category]

    FeedbackRequest:
      type: object
      properties:
        ingredient: { type: string }
        substitute:  { type: string }
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Rating from 1 (poor) to 5 (excellent)
      required: [ingredient, substitute, rating]

    FeedbackResponse:
      type: object
      properties:
        id:      { type: integer }
        message: { type: string }
      required: [id, message]