# lib/api-zod/src/generated/types/feedbackRequest.ts
/** VitalSub API – General-purpose ingredient substitution engine. */

export interface FeedbackRequest {
  ingredient: string;
  substitute: string;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  rating: number;
}