# lib/api-zod/src/generated/types/nutritionData.ts
/** VitalSub API – General-purpose ingredient substitution engine. */

export interface NutritionData {
  /** VitalSub API – General-purpose ingredient substitution engine. */
  calories?: number | null;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  protein?: number | null;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  carbs?: number | null;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  fat?: number | null;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  sugar?: number | null;
  /** VitalSub API – General-purpose ingredient substitution engine. */
  sodium?: number | null;
}